photo-server/app/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
numpy==1.26.3
sentence-transformers==2.3.1
python-multipart==0.0.6
aiofiles==23.2.1
jinja2==3.1.3
pillow==10.2.0
pillow-heif==0.15.0
